Fixed encoders, renamed steganogan.py to models.py in order to avoid conflicts with package name, removed utils folder.
