+++ steganogan/__init__.py
@@ -1,12 +1,14 @@
 # -*- coding: utf-8 -*-
 """Top-level package for SteganoGAN."""
 
-from steganogan.steganogan import Steganographer
-
 __author__ = """MIT Data To AI Lab"""
 __email__ = 'dailabmit@gmail.com'
 __version__ = '0.1.0-dev'
 
-from steganogan.steganogan import SteganoGAN
+from steganogan.critics import BasicCritic
+from steganogan.decoders import BasicDecoder
+from steganogan.encoders import BasicEncoder, DenseEncoder, ResidualEncoder
+from steganogan.models import SteganoGAN
 
-__all__ = ('SteganoGAN', )
+__all__ = ('BasicCritic', 'BasicDecoder', 'BasicEncoder',
+           'DenseEncoder', 'ResidualEncoder', 'SteganoGAN')
