--- steganogan/architectures/__init__.py
@@ -1,5 +1,0 @@
-from steganogan.architectures.critics import BasicCritic
-from steganogan.architectures.decoders import BasicDecoder
-from steganogan.architectures.encoders import BasicEncoder, DenseEncoder, ResidualEncoder
-
-__all__ = ('BasicCritic', 'BasicDecoder', 'BasicEncoder', 'DenseEncoder', 'ResidualEncoder')
